- Created hdf5compiler package structure - Added hdf5compile for moke
